delete: check for unresolved imports

+++ src/delete.ts
@@ -5,7 +5,15 @@ import { deleteFile } from './fs';
 export async function deleteUnimportedFiles(
   result: ProcessedResult,
   context: Context,
-): Promise<{ numFilesDeleted: number }> {
+): Promise<{ numFilesDeleted: number; error?: string }> {
+  const deleteIsSafe = result.unresolved.length === 0;
+  if (!deleteIsSafe) {
+    return {
+      numFilesDeleted: 0,
+      error:
+        'Unable to safely delete files while there are unresolved imports.',
+    };
+  }
   await Promise.all(
     result.unimported.map((file) => deleteFile(file, context.cwd)),
   );
